Split policy authoring files for CI limits
