+++ crates/tandem-server/src/app/state/app_state_impl_parts/part06.rs
@@ -13,7 +13,7 @@ fn automation_v2_definition_is_context_recovered(automation: &AutomationV2Spec) 
 /// Move an unparseable incident-monitor state file aside instead of silently
 /// discarding it. Loads run at startup with their errors ignored, so without
 /// this a corrupt file would be replaced by an empty default on the next
-/// persist — losing publish receipts / idempotency keys and re-filing duplicate
+/// persist ÃƒÂ¢Ã¢â€šÂ¬Ã¢â‚¬Â losing publish receipts / idempotency keys and re-filing duplicate
 /// external issues. Quarantining preserves the original bytes for recovery and
 /// logs loudly, while the caller continues with empty in-memory state.
 async fn quarantine_corrupt_incident_monitor_state(
@@ -193,48 +193,6 @@ fn policy_decision_permission(decision: &PolicyDecisionRecord) -> Option<AccessP
     })
     .or_else(|| decision.tool.as_ref().map(|_| AccessPermission::Execute))
     .or_else(|| decision.resource.as_ref().map(|_| AccessPermission::Read))
-}
-
-fn policy_decision_input_base(decision: &PolicyDecisionRecord) -> EnterprisePolicyInput {
-    let mut input = EnterprisePolicyInput::new(decision.tenant_context.clone());
-    if let Some(org_unit_id) = policy_decision_org_unit_id(decision) {
-        input = input.with_org_unit_id(org_unit_id);
-    }
-    if let Some(resource) = decision.resource.clone() {
-        input = input.with_resource(resource);
-    }
-    if let Some(workflow_id) = policy_decision_workflow_id(decision) {
-        input = input.with_workflow_id(workflow_id);
-    }
-    if let Some(workflow_phase) = policy_decision_workflow_phase(decision) {
-        input = input.with_workflow_phase(workflow_phase);
-    }
-    if let Some(permission) = policy_decision_permission(decision) {
-        input = input.with_permission(permission);
-    }
-    if let Some(tool) = decision.tool.clone() {
-        input = input.with_tool(tool);
-    }
-    if let Some(arguments) = ["/tool_arguments", "/arguments", "/tool/arguments"]
-        .iter()
-        .find_map(|pointer| decision.metadata.pointer(pointer).cloned())
-    {
-        input = input.with_arguments(arguments);
-    }
-    input
-}
-
-fn policy_decision_inputs(decision: &PolicyDecisionRecord) -> Vec<EnterprisePolicyInput> {
-    let input = policy_decision_input_base(decision);
-    if decision.data_classes.is_empty() {
-        return vec![input];
-    }
-    decision
-        .data_classes
-        .iter()
-        .copied()
-        .map(|data_class| input.clone().with_data_class(data_class))
-        .collect()
 }
 
 fn enterprise_policy_effect_priority(effect: EnterprisePolicyEffect) -> u8 {
@@ -464,7 +422,7 @@ impl AppState {
     }
 
     /// Persist migrated legacy state to the canonical path (one-time on read)
-    /// and log a deprecation warning. Failures are logged, not fatal — the
+    /// and log a deprecation warning. Failures are logged, not fatal ÃƒÂ¢Ã¢â€šÂ¬Ã¢â‚¬Â the
     /// in-memory state is already loaded and the legacy file is left intact.
     async fn migrate_legacy_incident_monitor_state<F, Fut>(
         &self,
@@ -967,33 +925,6 @@ impl AppState {
         Ok(())
     }
 
-    pub async fn persist_enterprise_policy_rules(&self) -> anyhow::Result<()> {
-        let records = {
-            let guard = self.enterprise.policy_rules.read().await;
-            guard
-                .iter()
-                .map(|(rule_id, rule)| {
-                    let tenant_context = rule
-                        .tenant_context
-                        .clone()
-                        .unwrap_or_else(tandem_enterprise_contract::TenantContext::local_implicit);
-                    crate::governance_store::GovernanceStoreFile::PolicyRules.json_record(
-                        rule_id,
-                        rule,
-                        &tenant_context,
-                        rule.org_unit_id.as_deref(),
-                    )
-                })
-                .collect::<anyhow::Result<Vec<_>>>()?
-        };
-        crate::governance_store::for_state(self)
-            .write_json_records(
-                crate::governance_store::GovernanceStoreFile::PolicyRules,
-                &records,
-            )
-            .await
-    }
-
     pub async fn persist_external_actions(&self) -> anyhow::Result<()> {
         if let Some(parent) = self.external_actions_path.parent() {
             fs::create_dir_all(parent).await?;
@@ -1183,7 +1114,7 @@ impl AppState {
 
     /// TAN-546: stamp a publish receipt with the tenant/workspace of its draft
     /// so tenant-scoped assessment reports can filter receipts the same way they
-    /// filter incidents and audit events. Only fills gaps — an adapter that
+    /// filter incidents and audit events. Only fills gaps ÃƒÂ¢Ã¢â€šÂ¬Ã¢â‚¬Â an adapter that
     /// already set the fields wins, and single-tenant drafts leave them None.
     async fn stamp_incident_monitor_post_tenant(&self, post: &mut IncidentMonitorPostRecord) {
         if post.tenant_id.is_some() || post.workspace_id.is_some() {
@@ -1423,27 +1354,6 @@ impl AppState {
                 )
             }),
         )
-    }
-
-    pub async fn resolve_enterprise_policy_input(
-        &self,
-        input: &EnterprisePolicyInput,
-        now_ms: u64,
-    ) -> anyhow::Result<tandem_enterprise_contract::EffectivePolicySnapshot> {
-        self.load_enterprise_policy_rules_if_needed().await?;
-        let rules = self
-            .enterprise
-            .policy_rules
-            .read()
-            .await
-            .values()
-            .cloned()
-            .collect::<Vec<_>>();
-        Ok(EnterprisePolicyResolver::new(rules).resolve(input, now_ms))
-    }
-
-    pub async fn ensure_enterprise_policy_rules_loaded(&self) -> anyhow::Result<()> {
-        self.load_enterprise_policy_rules_if_needed().await
     }
 
     async fn load_enterprise_policy_rules_if_needed(&self) -> anyhow::Result<()> {
